Split report.go
